# docs/openapi/v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tico API (v2)",
    "version": "2.0.0",
    "description": "The stable API for building your own frontend on Tico. Everything here keeps its shape within v2: fields are added, never removed or renamed, and a breaking change is a new /api/v3. Routes not listed are internal. Every write needs an `Idempotency-Key` header (1-200 characters; a retry with the same key and body returns the first answer). Errors are `{\"error\": {code, detail, retryable}}`. See docs/custom-frontend.md."
  },
...
  "paths": {
...
    "/api/v2/context/document": {
      "get": {
        "summary": "One document",
        "operationId": "getDocument",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "A problem: {\"error\": {\"code\", \"detail\", \"retryable\", ...}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request did not validate: {\"detail\": [{\"loc\", \"msg\", \"type\"}]}, or a Problem for a rule the server enforces"
          }
        },
        "tags": [
          "Docs"
        ]
      }
    },
    "/api/v2/context/search": {
      "get": {
        "summary": "Search company documents",
        "operationId": "searchDocs",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "title": "Q"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "docs",
                "market"
              ],
              "type": "string",
              "default": "all",
              "title": "Source"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocSearch"
                }
              }
            }
          },
          "401": {
            "description": "A problem: {\"error\": {\"code\", \"detail\", \"retryable\", ...}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request did not validate: {\"detail\": [{\"loc\", \"msg\", \"type\"}]}, or a Problem for a rule the server enforces"
          }
        },
        "tags": [
          "Docs"
        ]
      }
    },
...
    "/api/v2/docs/search": {
      "get": {
        "summary": "Search internal docs (ranked with bm25, with an excerpt) and linked docs (title, note, address) together; each result has a type",
        "operationId": "searchInternalAndLinkedDocs",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocsSearchResults"
                }
              }
            }
          },
          "401": {
            "description": "A problem: {\"error\": {\"code\", \"detail\", \"retryable\", ...}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request did not validate: {\"detail\": [{\"loc\", \"msg\", \"type\"}]}, or a Problem for a rule the server enforces"
          }
        },
        "tags": [
          "Docs"
        ]
      }
    },
...
    "/api/v2/meetings/search": {
      "get": {
        "summary": "Search or list recorded meetings",
        "operationId": "searchMeetings",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "person",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Person"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingSearch"
                }
              }
            }
          },
          "401": {
            "description": "A problem: {\"error\": {\"code\", \"detail\", \"retryable\", ...}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request did not validate: {\"detail\": [{\"loc\", \"msg\", \"type\"}]}, or a Problem for a rule the server enforces"
          }
        },
        "tags": [
          "Meetings"
        ]
      }
    },
    "/api/v2/meetings/transcript": {
      "get": {
        "summary": "A meeting's transcript",
        "operationId": "getMeetingTranscript",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 1,
              "default": 20000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "A problem: {\"error\": {\"code\", \"detail\", \"retryable\", ...}}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request did not validate: {\"detail\": [{\"loc\", \"msg\", \"type\"}]}, or a Problem for a rule the server enforces"
          }
        },
        "tags": [
          "Meetings"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocSearch": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string"
          },
          "results": {
            "type": "array"
          },
          "has_more": {
            "type": "boolean"
          },
          "mode": {
            "type": "string"
          }
        },
        "required": [
          "has_more",
          "mode",
          "query",
          "results"
        ],
        "additionalProperties": true
      },
      "DocsSearchResults": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "type": {
                      "enum": [
                        "internal"
                      ]
                    },
                    "id": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "excerpt": {
                      "type": "string"
                    },
                    "score": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "excerpt",
                    "id",
                    "path",
                    "score",
                    "title",
                    "type"
                  ],
                  "additionalProperties": true
                },
                {
                  "type": "object",
                  "properties": {
                    "type": {
                      "enum": [
                        "linked"
                      ]
                    },
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "score": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "description",
                    "id",
                    "kind",
                    "score",
                    "title",
                    "type",
                    "url"
                  ],
                  "additionalProperties": true
                }
              ]
            }
          }
        },
        "required": [
          "results"
        ],
        "additionalProperties": true
      },
...
      "MeetingSearch": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array"
          },
          "next_offset": {
            "type": [
              "integer",
              "null"
            ]
          },
          "mode": {
            "type": "string"
          }
        },
        "required": [
          "mode",
          "next_offset",
          "results"
        ],
        "additionalProperties": true
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "detail"
            ],
            "additionalProperties": true,
            "properties": {
              "code": {
                "type": "string"
              },
              "detail": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "sign_in": {
                "type": "string",
                "description": "On a 401 with built-in sign-in: the login path"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "A bearer session from POST /auth/token (browser apps) or a personal API token (servers)."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tico_session",
        "description": "The browser session of Tico's own page (`__Host-tico_session` over https)."
      }
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}